no doubles
slightly improved

+++ Structures.h
@@ -5,8 +5,8 @@
 
 
 struct position {
-    float x{};
-    float y{};
+    int x{};
+    int y{};
 };
 
 struct IParam {
